+++ .github/CHANGELOG.md
@@ -108,6 +108,10 @@
   that performed post-processing on a vector-valued QNode.
   [(#)](https://github.com/PennyLaneAI/pennylane/pull/)
 
+* Fixed a bug in the initialization of `QubitUnitary` where the size of
+  the matrix was not checked against the number of wires.
+  [(#1439)](https://github.com/PennyLaneAI/pennylane/pull/1439)
+
 <h3>Documentation</h3>
 
 <h3>Contributors</h3>